new delete creatures and reset properties buttons + start of implementation of link creature_list and groupbox
Implementation of button_delete_all which erases all created creatures when pressed. Implementation of button_reset, which sets all properties to 0 or to unchecked when clicked. Moved all properties (sliders etc.) to a group box so that it is easier and quicker to show all or hide all at once. Beginning of implementation of a link from creature_list to group box visibility, but not yet differentiating the cases (i.e. one different group box of properties for each creature (+ diff group boxes as diff properties according creature (prey/plant/pred) ).