Display features of the Simulation button (simBut)
I wanted this button to show only if some creatures existed. Therefore i added to the functions to create these creatures (linked with the "add omnivorous creature", "add herbivore creature" and "add plant" buttons) or delete them (linked with the "delete all" and "delete creature" buttons), some conditions on the visibility of this simulation button.